requirements.txt
# Core dependencies
python-dotenv==1.1.1
requests==2.31.0

# LangChain (compatible versions)
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.11
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchainhub==0.1.21

# CrewAI
crewai==0.203.1
crewai-tools==0.76.0

# Data processing
pandas==2.2.2
Faker==25.3.0

# Google API
google-api-python-client==2.143.0

# Web framework
flask==3.1.2

# Production server
gunicorn==23.0.0
